Remove CLI related changes
